Updated Add Button
Add button will now check to make sure flash drive is in first before prompting user. It then adds file names to a text file. Checks if the file is in the flash drive already. If it isn't then it gets added to the flash drive. Added method searchFlashDrive to search flash drive for a file that is passed to it and returns a Boolean to indicate if file is present in the flash drive or not. Need to fix the Add Button method to not allow for the same file to be added to the text file. The same file could be added to the text file multiple times. However it won't add the file to the flash drive multiple times.